Refactor: Split loss function file into smaller modules
